src/requirements.txt
click==7.1.2
mysql-connector-python==8.0.23
protobuf==3.15.8
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
python-dotenv==0.17.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
six==1.15.0
bcrypt==3.2.0
pytest==6.2.3
pytest-tornasync==0.6.0.post2